requirements.txt
langchain-core
langchain-openai
python-dotenv
fastapi
uvicorn
langserve
sse_starlette